requirements.txt
matplotlib==3.1.2
numpy==1.18.0
pandas==0.25.3
PyYAML==5.2
requests==2.22.0

